interpreter: Split exception calsses from interpreterbase.py
